P04 - First Pass, Program
Now the program makes an output file (output.txt) and the output goes
there. It calculates the area of the pool and the time it would take our
workers to paint it. There may be some errors I didn't catch though,
since I didn't spend too much time testing this.